extract text creation dialog
